added unsigned to release date and to sale in the albums table

+++ albums_migration.sql
@@ -5,8 +5,8 @@ create table if not exists albums
     id INT unsigned not null auto_increment,
     artist varchar(255) not null,
     name varchar(255) not null,
-    release_date INT not null,
-    sales float not null,
+    release_date INT unsigned not null,
+    sales float unsigned not null,
     genre varchar(255) not null,
      PRIMARY KEY (id)
